server/requirements.txt
flask
flask-cors
ddgs
beautifulsoup4
requests
pandas
playwright
lxml
youtube-transcript-api
arxiv
feedparser
python-dotenv
rank_bm25
sentence-transformers
numpy
scikit-learn
torch
tf-keras
openai
newspaper3k
langchain-community
langchain-huggingface
langchain-openai
langchain-text-splitters
chromadb
trafilatura

